feature/likes 추가 삭제 기능 (수정4)

+++ services/likeService.js
@@ -1,10 +1,27 @@
 const likeDao = require('../models/likeDao');
+const productDao = require('../models/productDao');
 
 const createLike = async (userId, productId) => {
+  const product = await productDao.getProductById(productId);
+
+  if (!product) {
+    const error = new Error('Product_Does_Not_Exist');
+    error.statusCode = 404;
+    throw error;
+  }
+
   return await likeDao.createLike(userId, productId);
 };
 
 const deleteLike = async (userId, productId) => {
+  const product = await productDao.getProductById(productId);
+
+  if (!product) {
+    const error = new Error('Product_Does_Not_Exist');
+    error.statusCode = 404;
+    throw error;
+  }
+
   return await likeDao.deleteLike(userId, productId);
 };
